Divided ui.c up into component files.
