+++ openrtx/include/ui/ui_default.h
@@ -107,30 +107,6 @@ enum
     GUI_LINE_NUM_OF
 };
 
-// GUI Commands
-enum
-{
-    GUI_CMD_NULL         = 0x00 ,
-    GUI_CMD_EVENT_START  = 0x01 ,
-    GUI_CMD_EVENT_END    = 0x02 ,
-    GUI_CMD_LINE_STYLE   = 0x03 ,
-    GUI_CMD_ALIGN_LEFT   = 0x04 ,
-    GUI_CMD_ALIGN_CENTER = 0x05 ,
-    GUI_CMD_ALIGN_RIGHT  = 0x06 ,
-    GUI_CMD_FONT_SIZE    = 0x07 ,
-    GUI_CMD_LINE_END     = 0x0A ,
-    GUI_CMD_LINK         = 0x0B ,
-    GUI_CMD_LINK_END     = 0x0C ,
-    GUI_CMD_PAGE         = 0x0E ,
-    GUI_CMD_VALUE        = 0x0F ,
-    GUI_CMD_TITLE        = 0x10 ,
-    GUI_CMD_TEXT         = 0x11 ,
-    GUI_CMD_STUBBED      = 0x1E ,
-    GUI_CMD_PAGE_END     = 0x1F ,
-    GUI_CMD_DATA_AREA           ,
-    GUI_CMD_NUM_OF
-};
-
 enum
 {
     LINK_MAX_NUM_OF = 12
@@ -142,112 +118,6 @@ typedef enum
     LINK_TYPE_PAGE  ,
     LINK_TYPE_VALUE
 }LinkType_en;
-
-// GUI Values - Display
-enum
-{
-    GUI_VAL_CURRENT_TIME     ,
-    GUI_VAL_BATTERY_LEVEL    ,
-    GUI_VAL_LOCK_STATE       ,
-    GUI_VAL_MODE_INFO        ,
-    GUI_VAL_FREQUENCY        ,
-    GUI_VAL_RSSI_METER       ,
-
-    GUI_VAL_BANKS            ,
-    GUI_VAL_CHANNELS         ,
-    GUI_VAL_CONTACTS         ,
-    GUI_VAL_GPS              ,
-    // Settings
-    // Display
-#ifdef SCREEN_BRIGHTNESS
-    GUI_VAL_BRIGHTNESS       ,
-#endif // SCREEN_BRIGHTNESS
-#ifdef SCREEN_CONTRAST
-    GUI_VAL_CONTRAST         ,
-#endif // SCREEN_CONTRAST
-    GUI_VAL_TIMER            ,
-    // Time and Date
-    GUI_VAL_DATE             ,
-    GUI_VAL_TIME             ,
-    // GPS
-    GUI_VAL_GPS_ENABLED      ,
-    GUI_VAL_GPS_SET_TIME     ,
-    GUI_VAL_GPS_TIME_ZONE    ,
-    // Radio
-    GUI_VAL_RADIO_OFFSET     ,
-    GUI_VAL_RADIO_DIRECTION  ,
-    GUI_VAL_RADIO_STEP       ,
-    // M17
-    GUI_VAL_M17_CALLSIGN     ,
-    GUI_VAL_M17_CAN          ,
-    GUI_VAL_M17_CAN_RX_CHECK ,
-    // Accessibility - Voice
-    GUI_VAL_LEVEL            ,
-    GUI_VAL_PHONETIC         ,
-    // Info
-    GUI_VAL_BATTERY_VOLTAGE  ,
-    GUI_VAL_BATTERY_CHARGE   ,
-    GUI_VAL_RSSI             ,
-    GUI_VAL_USED_HEAP        ,
-    GUI_VAL_BAND             ,
-    GUI_VAL_VHF              ,
-    GUI_VAL_UHF              ,
-    GUI_VAL_HW_VERSION       ,
-#ifdef PLATFORM_TTWRPLUS
-    GUI_VAL_RADIO            ,
-    GUI_VAL_RADIO_FW         ,
-#endif // PLATFORM_TTWRPLUS
-    GUI_VAL_STUBBED          ,
-    GUI_VAL_NUM_OF
-};
-
-// GUI Values - Input
-enum
-{
-    GUI_VAL_INPUT_VFO_MIDDLE ,
-
-    GUI_VALUE_INPUT_NUM_OF = 12
-};
-
-typedef enum
-{
-#ifdef SCREEN_BRIGHTNESS
-    D_BRIGHTNESS ,          // GUI_VAL_INP_BRIGHTNESS
-#endif
-#ifdef SCREEN_CONTRAST
-    D_CONTRAST   ,          // GUI_VAL_INP_CONTRAST
-#endif
-    D_TIMER                 // GUI_VAL_INP_TIMER
-}DisplayItems_en;
-
-#ifdef GPS_PRESENT
-typedef enum
-{
-    G_ENABLED  ,            // GUI_VAL_INP_ENABLED
-    G_SET_TIME ,            // GUI_VAL_INP_SET_TIME
-    G_TIMEZONE              // GUI_VAL_INP_TIMEZONE
-}SettingsGPSItems_en;
-#endif
-
-typedef enum
-{
-    VP_LEVEL    ,           // GUI_VAL_INP_LEVEL
-    VP_PHONETIC             // GUI_VAL_INP_PHONETIC
-}SettingsVoicePromptItems_en;
-
-typedef enum
-{
-    R_OFFSET    ,           // GUI_VAL_INP_OFFSET
-    R_DIRECTION ,           // GUI_VAL_INP_DIRECTION
-    R_STEP                  // GUI_VAL_INP_STEP
-}SettingsRadioItems_en;
-
-typedef enum
-{
-    M17_CALLSIGN ,          // GUI_VAL_INP_CALLSIGN
-    M17_CAN      ,          // GUI_VAL_INP_CAN
-    M17_CAN_RX              // GUI_VAL_INP_CAN_RX
-}SettingsM17Items_en;
 
 typedef struct
 {
@@ -377,83 +247,6 @@ typedef struct
 extern GuiState_st  GuiState ;
 extern State_st     last_state ;
 extern bool         macro_latched ;
-
-// Color Variable Field Unshifted Masks and Shifts
-enum
-{
-    COLOR_ENC_RED_SHIFT   =   24 ,
-    COLOR_ENC_RED_MASK    = 0xFF ,
-    COLOR_ENC_GREEN_SHIFT =   16 ,
-    COLOR_ENC_GREEN_MASK  = 0xFF ,
-    COLOR_ENC_BLUE_SHIFT  =    8 ,
-    COLOR_ENC_BLUE_MASK   = 0xFF ,
-    COLOR_ENC_ALPHA_SHIFT =    0 ,
-    COLOR_ENC_ALPHA_MASK  = 0xFF
-};
-
-// Color Variable Field Encoder
-#define COLOR_DEF_ENC( r , g , b , a ) (const uint32_t)(                        \
-    ( ( (const uint32_t)r & COLOR_ENC_RED_MASK   ) << COLOR_ENC_RED_SHIFT   ) | \
-    ( ( (const uint32_t)g & COLOR_ENC_GREEN_MASK ) << COLOR_ENC_GREEN_SHIFT ) | \
-    ( ( (const uint32_t)b & COLOR_ENC_BLUE_MASK  ) << COLOR_ENC_BLUE_SHIFT  ) | \
-    ( ( (const uint32_t)a & COLOR_ENC_ALPHA_MASK ) << COLOR_ENC_ALPHA_SHIFT )   )
-
-// Color Definitions
-#define COLOR_DEF_WHITE         COLOR_DEF_ENC( 255 , 255 , 255 , 255 )
-#define COLOR_DEF_BLACK         COLOR_DEF_ENC(   0 ,   0 ,  0  , 255 )
-#define COLOR_DEF_GREY          COLOR_DEF_ENC(  60 ,  60 ,  60 , 255 )
-#define COLOR_DEF_ALPHA_GREY    COLOR_DEF_ENC(   0 ,   0 ,   0 , 255 )
-#define COLOR_DEF_YELLOW_FAB413 COLOR_DEF_ENC( 250 , 180 ,  19 , 255 )
-#define COLOR_DEF_RED           COLOR_DEF_ENC( 255 ,   0 ,   0 , 255 )
-#define COLOR_DEF_GREEN         COLOR_DEF_ENC(   0 , 255 ,   0 , 255 )
-#define COLOR_DEF_BLUE          COLOR_DEF_ENC(   0 ,   0 , 255 , 255 )
-
-// Color Table Definition
-#define COLOR_TABLE             \
-    COLOR_DEF_BLACK         ,   \
-    COLOR_DEF_WHITE         ,   \
-    COLOR_DEF_GREY          ,   \
-    COLOR_DEF_ALPHA_GREY    ,   \
-    COLOR_DEF_YELLOW_FAB413 ,   \
-    COLOR_DEF_RED           ,   \
-    COLOR_DEF_GREEN         ,   \
-    COLOR_DEF_BLUE
-
-// Color Selectors
-typedef enum
-{
-    COLOR_BLACK         ,
-    COLOR_WHITE         ,
-    COLOR_GREY          ,
-    COLOR_ALPHA_GREY    ,
-    COLOR_YELLOW_FAB413 ,
-    COLOR_RED           ,
-    COLOR_GREEN         ,
-    COLOR_BLUE          ,
-    COLOR_NUM_OF        ,
-
-    COLOR_BG  = COLOR_BLACK         , // background
-    COLOR_FG  = COLOR_WHITE         , // foreground
-    COLOR_GG  = COLOR_GREY          , // grey ground
-    COLOR_AGG = COLOR_ALPHA_GREY    , // alpha grey ground
-    COLOR_HL  = COLOR_YELLOW_FAB413 , // highlight
-    COLOR_OP0 = COLOR_RED           , // option 0
-    COLOR_OP1 = COLOR_GREEN         , // option 1
-    COLOR_OP2 = COLOR_BLUE          , // option 2
-    COLOR_OP3 = COLOR_YELLOW_FAB413 , // option 3
-    COLOR_AL  = COLOR_RED             // alarm
-
-}ColorSelector_en;
-
-// Load the Color_st structure with the color variable fields
-#define COLOR_LD( c , cv )                                                       \
-c->red   = (uint8_t)( ( cv >> COLOR_ENC_RED_SHIFT   ) & COLOR_ENC_RED_MASK   ) ; \
-c->green = (uint8_t)( ( cv >> COLOR_ENC_GREEN_SHIFT ) & COLOR_ENC_GREEN_MASK ) ; \
-c->blue  = (uint8_t)( ( cv >> COLOR_ENC_BLUE_SHIFT  ) & COLOR_ENC_BLUE_MASK  ) ; \
-c->alpha = (uint8_t)( ( cv >> COLOR_ENC_ALPHA_SHIFT ) & COLOR_ENC_ALPHA_MASK ) ;
-
-// Color load fn.
-extern void uiColorLoad( Color_st* color , ColorSelector_en colorSelector );
 
 extern const uiPageDesc_st* uiGetPageDesc( uiPageNum_en pageNum );
 extern const char**         uiGetPageLoc( uiPageNum_en pageNum );
