+++ openrtx/src/ui/default/ui_value_arrays.c
@@ -1,0 +1,77 @@
+/***************************************************************************
+ *   Copyright (C) 2020 - 2024 by Federico Amedeo Izzo IU2NUO,             *
+ *                                Niccolò Izzo IU2KIN                      *
+ *                                Frederik Saraci IU2NRO                   *
+ *                                Silvano Seva IU2KWO                      *
+ *                                Kim Lyon VK6KL                           *
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 3 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ *   This program is distributed in the hope that it will be useful,       *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU General Public License for more details.                          *
+ *                                                                         *
+ *   You should have received a copy of the GNU General Public License     *
+ *   along with this program; if not, see <http://www.gnu.org/licenses/>   *
+ ***************************************************************************/
+
+#include <ui/ui_default.h>
+#include "ui_value_arrays.h"
+
+const char* symbols_ITU_T_E161[] =
+{
+    " 0"        ,
+    ",.?1"      ,
+    "abc2ABC"   ,
+    "def3DEF"   ,
+    "ghi4GHI"   ,
+    "jkl5JKL"   ,
+    "mno6MNO"   ,
+    "pqrs7PQRS" ,
+    "tuv8TUV"   ,
+    "wxyz9WXYZ" ,
+    "-/*"       ,
+    "#"
+};
+
+const char* symbols_ITU_T_E161_callsign[] =
+{
+    "0 "    ,
+    "1"     ,
+    "ABC2"  ,
+    "DEF3"  ,
+    "GHI4"  ,
+    "JKL5"  ,
+    "MNO6"  ,
+    "PQRS7" ,
+    "TUV8"  ,
+    "WXYZ9" ,
+    "-/"    ,
+    ""
+};
+
+const char* display_timer_values[] =
+{
+    "Off" ,
+    "5 s" ,
+    "10 s" ,
+    "15 s" ,
+    "20 s" ,
+    "25 s" ,
+    "30 s" ,
+    "1 min" ,
+    "2 min" ,
+    "3 min" ,
+    "4 min" ,
+    "5 min" ,
+    "15 min" ,
+    "30 min" ,
+    "45 min" ,
+    "1 hour"
+};
+
+
