+++ tests/platform/gui_demo.c
@@ -44,7 +44,7 @@ int main(void)
     // Print splash screen
     Pos_st splash_origin = {0, SCREEN_HEIGHT / 2};
     Color_st color_op3 ;
-    uiColorLoad( &color_op3 , COLOR_OP3 );
+    ui_ColorLoad( &color_op3 , COLOR_OP3 );
     gfx_clearScreen();
     gfx_print(splash_origin, FONT_SIZE_4, ALIGN_CENTER,
               color_op3, "OpenRTX");
